Add feature adder for structured problem; separate general optimization params from problem type specific params (structured vs structureless) (#12)

+++ include/vslam_types.h
@@ -114,7 +114,7 @@ struct StructuredVisionFeatureTrack {
   /**
    * 3D coordinate of the feature tracked by the feature track.
    */
-  Eigen::Vector3f point;
+  Eigen::Vector3d point;
   /**
    * Image feature track - same as the structureless feature track.
    * */
@@ -126,7 +126,7 @@ struct StructuredVisionFeatureTrack {
   /**
    * Convenience constructor: initialize everything.
    */
-  StructuredVisionFeatureTrack(const Eigen::Vector3f& point,
+  StructuredVisionFeatureTrack(const Eigen::Vector3d& point,
                                const VisionFeatureTrack& feature_track)
       : point(point), feature_track(feature_track){};
 };
